+++ test/aws_encryption_sdk_test.exs
@@ -1,8 +1,7 @@
 defmodule AwsEncryptionSdkTest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
   doctest AwsEncryptionSdk
 
-  test "greets the world" do
-    assert AwsEncryptionSdk.hello() == :world
-  end
+  # Integration tests are in test/aws_encryption_sdk/integration_test.exs
+  # This file just runs doctests from the main module
 end